separating concerns
