+++ src/ConfigurationBinder/Extensions/ServiceCollectionExtensions.cs
@@ -33,63 +33,6 @@ namespace ConfigurationBinder.Extensions
             where TConfig : class, new() =>
             services.AddOptions<TConfig>()
                 .Configure<IConfiguration>((target, configuration) =>
-                    BindSettings(configuration, target, options));
-
-        public static void BindSettings(IConfiguration configuration, object target, ConfigurationBinderOptions options)
-        {
-            var type = target.GetType();
-            var settableProperties = type
-                .GetProperties()
-                .Where(prop => prop.GetSetMethod() != null);
-
-            foreach (var prop in settableProperties)
-            {
-                var key = $"{type.Name}.{prop.Name}";
-
-                string value = configuration
-                    .AsEnumerable()
-                    .FirstOrDefault(kv => kv.Key.Equals(key, options.KeyComparison))
-                    .Value;
-
-                object propertyValue = ParserFactory
-                    .CreateParser(prop.PropertyType, options)
-                    .Parse(value);
-
-                AssignProperty(prop, target, propertyValue);
-            }
-        }
-
-        public static void AssignProperty(PropertyInfo prop, object target, object value)
-        {
-            var propertyType = prop.PropertyType;
-
-            if (propertyType.IsEnumerable()
-                && propertyType != typeof(string)
-                && value is Array array)
-            {
-
-                var elementType = propertyType.IsGenericType 
-                    ? propertyType.GetGenericArguments()[0] 
-                    : propertyType.GetElementType();
-
-                var arrayType = elementType.MakeArrayType();
-
-                var creationArgs = new object[] { array.Length };
-
-                var targetArray = (Array)Activator.CreateInstance(arrayType, creationArgs);
-
-                for (int i = 0; i < array.Length; i++)
-                {
-                    var elementValue = Convert.ChangeType(array.GetValue(i), elementType);
-                    targetArray.SetValue(elementValue, i);
-                }
-
-                prop.SetValue(target, targetArray);
-            }
-            else
-            {
-                prop.SetValue(target, value);
-            }
-        }
+                    configuration.BindSettings(target, options));
     }
 }
